improve allocation of large storage arrays for each supported type
This used to be revision r1934.

+++ ij2-common/src/main/java/imagej/data/LongType.java
@@ -91,7 +91,7 @@ public class LongType implements Type
 	}
 
 	@Override
-	public Object allocateStorageArray(int numPixels)
+	public Object allocateStorageArray(long numPixels)
 	{
 		long numStorageUnits = calcNumStorageUnitsFromPixelCount(numPixels);
