+++ ij2-common/src/main/java/imagej/data/UnsignedIntType.java
@@ -91,7 +91,7 @@ public class UnsignedIntType implements Type
 	}
 
 	@Override
-	public Object allocateStorageArray(int numPixels)
+	public Object allocateStorageArray(long numPixels)
 	{
 		long numStorageUnits = calcNumStorageUnitsFromPixelCount(numPixels);
